+++ common/src/main/java/de/maxhenkel/voicechat/gui/CreateGroupScreen.java
@@ -44,11 +44,11 @@ public class CreateGroupScreen extends VoiceChatScreenBase {
         Keyboard.enableRepeatEvents(true);
 
         groupName = new GuiTextField(0, fontRenderer, guiLeft + 7, guiTop + 32, xSize - 7 * 2, 10);
-        groupName.setMaxStringLength(16);
+        groupName.setMaxStringLength(24);
         groupName.setValidator(s -> s.isEmpty() || Voicechat.GROUP_REGEX.matcher(s).matches());
 
         password = new GuiTextField(1, fontRenderer, guiLeft + 7, guiTop + 58, xSize - 7 * 2, 10);
-        password.setMaxStringLength(16);
+        password.setMaxStringLength(32);
         password.setValidator(s -> s.isEmpty() || Voicechat.GROUP_REGEX.matcher(s).matches());
 
         groupTypeButton = new ButtonBase(2, guiLeft + 6, guiTop + 71, xSize - 12, 20, GROUP_TYPE.getUnformattedComponentText() + ": " + groupType.getTranslation().getUnformattedComponentText()) {